+++ pkg/linter/lib/src/rules/unreachable_from_main.dart
@@ -610,24 +610,19 @@ extension on LibraryElement {
 }
 
 extension on Element {
-  bool get hasVisibleForTesting => switch (this) {
-    Annotatable(:var metadata) => metadata.hasVisibleForTesting,
-    _ => false,
-  };
-  bool get hasWidgetPreview => switch (this) {
-    Annotatable(:var metadata) =>
+  bool get hasVisibleForTesting => metadata.hasVisibleForTesting;
+
+  bool get hasWidgetPreview =>
       // Widget previews can be applied to public:
       //   - Constructors (generative and factory)
       //   - Top-level functions
       //   - Static member functions
       (this is ConstructorElement ||
-              this is TopLevelFunctionElement ||
-              (this is ExecutableElement &&
-                  (this as ExecutableElement).isStatic)) &&
-          !isPrivate &&
-          metadata.hasWidgetPreview,
-    _ => false,
-  };
+          this is TopLevelFunctionElement ||
+          (this is ExecutableElement &&
+              (this as ExecutableElement).isStatic)) &&
+      !isPrivate &&
+      metadata.hasWidgetPreview;
   bool get isPragma => (library?.isDartCore ?? false) && name == 'pragma';
   bool get isWidgetPreview =>
       (library?.isWidgetPreviews ?? false) && name == 'Preview';